IEquipable now inherits IMultiResource.
This simplifies inheritance by removing to classes: 1 interface and 1 abstract.
To do this, IEquippable uses the same resource and stores additional data in simple mappings.
We call the Resource + base address + equipableRefId: ExtendedResource.
ExtendedResource is only used for input and output.
Utility functions are added to addResourceEntry for a ExtendedResource and equivalent for all the getters.