Refactoring of approvals (#39)
Resources now have independent approvals which apply to accepting, rejecting and setting priorities on resources.
On Equippables: equip, unequip and replace equip use the nesting permissions. For this a new method was added to the IRMRKNestingWithEquippable interface: isApprovedOrOwner so permissions can be fully checked with a single external call.
Tests for added for nesting, resources and resources on equippable for both approvals and approvals for all.

* Splits approvals for Resources from token management.

* Adds tests for approval for all on multi resources.

* Renames modifier and adds FIXME.

* Renables tests for approvals on equippable resources tests.

* Equippable now allows operations for approved addresses.

* Adds approval tests for RMRKEquippable.

Equip, unequip, and replace equipement can be done if sender is approved for token or for all.

* Adds test for approvals in nesting

Approved are allowed to accept, reject, remove and unnest self.